refactor: extract admin API key handlers to layered architecture
- Create api/admin/ module directory
- Create api/admin/api_keys.rs
- Create repositories/api_key_repository.rs
- Remove 5 SQL functions from db/queries.rs
- Remove 5 handlers from router.rs
