+++ src/api/mod.rs
@@ -1,3 +1,4 @@
+pub mod admin;
 pub mod analytics;
 pub mod auth;
 pub mod billing;
