+++ src/db/queries.rs
@@ -2894,217 +2894,6 @@ pub struct AdminApiKeyRecord {
     pub updated_by: Option<String>,
 }
 
-pub async fn list_admin_api_keys(
-    db: &worker::d1::D1Database,
-    page: i64,
-    limit: i64,
-    search: Option<&str>,
-    status_filter: Option<&str>,
-) -> worker::Result<(Vec<AdminApiKeyRecord>, i64)> {
-    let offset = (page - 1) * limit;
-
-    // Build status filter clause
-    let status_clause = match status_filter {
-        Some("active") => "AND ak.status = 'active'",
-        Some("revoked") => "AND ak.status = 'revoked'",
-        Some("deleted") => "AND ak.status = 'deleted'",
-        _ => "", // 'all' or None - no filter
-    };
-
-    if let Some(s) = search {
-        let pattern = format!("%{}%", s);
-        let where_clause = format!(
-            "WHERE (u.email LIKE ?1 OR ak.name LIKE ?1 OR o.name LIKE ?1) {}",
-            status_clause
-        );
-
-        let count_sql = format!(
-            "SELECT COUNT(*) as cnt
-             FROM api_keys ak
-             JOIN users u ON ak.user_id = u.id
-             JOIN organizations o ON ak.org_id = o.id
-             LEFT JOIN billing_accounts ba ON o.billing_account_id = ba.id
-             {}",
-            where_clause
-        );
-
-        let list_sql = format!(
-            "SELECT ak.id, ak.name, ak.hint, ak.user_id,
-                    u.email as user_email, u.name as user_name,
-                    ak.org_id, o.name as org_name,
-                    ba.tier,
-                    ak.created_at, ak.last_used_at, ak.expires_at,
-                    ak.status, ak.updated_at, ak.updated_by
-             FROM api_keys ak
-             JOIN users u ON ak.user_id = u.id
-             JOIN organizations o ON ak.org_id = o.id
-             LEFT JOIN billing_accounts ba ON o.billing_account_id = ba.id
-             {}
-             ORDER BY ak.created_at DESC
-             LIMIT ?2 OFFSET ?3",
-            where_clause
-        );
-
-        #[derive(serde::Deserialize)]
-        struct CountRow {
-            cnt: i64,
-        }
-        let total = db
-            .prepare(&count_sql)
-            .bind(&[pattern.clone().into()])?
-            .first::<CountRow>(None)
-            .await?
-            .map(|r| r.cnt)
-            .unwrap_or(0);
-
-        let keys = db
-            .prepare(&list_sql)
-            .bind(&[
-                pattern.clone().into(),
-                (limit as f64).into(),
-                (offset as f64).into(),
-            ])?
-            .all()
-            .await?
-            .results::<AdminApiKeyRecord>()?;
-
-        Ok((keys, total))
-    } else {
-        let where_clause = if status_clause.is_empty() {
-            "".to_string()
-        } else {
-            format!("WHERE {}", status_clause.trim_start_matches("AND "))
-        };
-
-        let count_sql = format!(
-            "SELECT COUNT(*) as cnt
-             FROM api_keys ak
-             JOIN users u ON ak.user_id = u.id
-             JOIN organizations o ON ak.org_id = o.id
-             LEFT JOIN billing_accounts ba ON o.billing_account_id = ba.id
-             {}",
-            where_clause
-        );
-
-        let list_sql = format!(
-            "SELECT ak.id, ak.name, ak.hint, ak.user_id,
-                    u.email as user_email, u.name as user_name,
-                    ak.org_id, o.name as org_name,
-                    ba.tier,
-                    ak.created_at, ak.last_used_at, ak.expires_at,
-                    ak.status, ak.updated_at, ak.updated_by
-             FROM api_keys ak
-             JOIN users u ON ak.user_id = u.id
-             JOIN organizations o ON ak.org_id = o.id
-             LEFT JOIN billing_accounts ba ON o.billing_account_id = ba.id
-             {}
-             ORDER BY ak.created_at DESC
-             LIMIT ?1 OFFSET ?2",
-            where_clause
-        );
-
-        #[derive(serde::Deserialize)]
-        struct CountRow {
-            cnt: i64,
-        }
-        let total = db
-            .prepare(&count_sql)
-            .first::<CountRow>(None)
-            .await?
-            .map(|r| r.cnt)
-            .unwrap_or(0);
-
-        let keys = db
-            .prepare(&list_sql)
-            .bind(&[(limit as f64).into(), (offset as f64).into()])?
-            .all()
-            .await?
-            .results::<AdminApiKeyRecord>()?;
-
-        Ok((keys, total))
-    }
-}
-
-pub async fn admin_revoke_api_key(
-    db: &worker::d1::D1Database,
-    key_id: &str,
-    revoked_by_user_id: &str,
-    timestamp: i64,
-) -> worker::Result<bool> {
-    let stmt = db.prepare(
-        "UPDATE api_keys SET status = 'revoked', updated_at = ?1, updated_by = ?2
-         WHERE id = ?3 AND status = 'active'",
-    );
-    stmt.bind(&[
-        (timestamp as f64).into(),
-        revoked_by_user_id.into(),
-        key_id.into(),
-    ])?
-    .run()
-    .await?;
-    Ok(true)
-}
-
-pub async fn admin_reactivate_api_key(
-    db: &worker::d1::D1Database,
-    key_id: &str,
-    reactivated_by_user_id: &str,
-    timestamp: i64,
-) -> worker::Result<bool> {
-    let stmt = db.prepare(
-        "UPDATE api_keys SET status = 'active', updated_at = ?1, updated_by = ?2
-         WHERE id = ?3 AND status = 'revoked'",
-    );
-    stmt.bind(&[
-        (timestamp as f64).into(),
-        reactivated_by_user_id.into(),
-        key_id.into(),
-    ])?
-    .run()
-    .await?;
-    Ok(true)
-}
-
-pub async fn admin_delete_api_key(
-    db: &worker::d1::D1Database,
-    key_id: &str,
-    deleted_by_user_id: &str,
-    timestamp: i64,
-) -> worker::Result<bool> {
-    let stmt = db.prepare(
-        "UPDATE api_keys SET status = 'deleted', updated_at = ?1, updated_by = ?2
-         WHERE id = ?3 AND status != 'deleted'",
-    );
-    stmt.bind(&[
-        (timestamp as f64).into(),
-        deleted_by_user_id.into(),
-        key_id.into(),
-    ])?
-    .run()
-    .await?;
-    Ok(true)
-}
-
-pub async fn admin_restore_api_key(
-    db: &worker::d1::D1Database,
-    key_id: &str,
-    restored_by_user_id: &str,
-    timestamp: i64,
-) -> worker::Result<bool> {
-    let stmt = db.prepare(
-        "UPDATE api_keys SET status = 'active', updated_at = ?1, updated_by = ?2
-         WHERE id = ?3 AND status = 'deleted'",
-    );
-    stmt.bind(&[
-        (timestamp as f64).into(),
-        restored_by_user_id.into(),
-        key_id.into(),
-    ])?
-    .run()
-    .await?;
-    Ok(true)
-}
-
 #[cfg(test)]
 mod api_key_tests {
     #[test]
