+++ src/repositories/mod.rs
@@ -15,6 +15,7 @@
 
 // Add repository modules here as they are created:
 pub mod analytics_repository;
+pub mod api_key_repository;
 pub mod billing_repository;
 pub mod product_repository;
 pub mod report_repository;
@@ -22,6 +23,7 @@ pub mod settings_repository;
 pub mod tag_repository;
 pub mod user_repository;
 pub use analytics_repository::AnalyticsRepository;
+pub use api_key_repository::ApiKeyRepository;
 pub use billing_repository::BillingRepository;
 pub use product_repository::ProductRepository;
 pub use report_repository::ReportRepository;
